requirements.txt
asgiref==3.7.2
Django==5.0.1
django-ckeditor-5==0.2.11
django-cors-headers==4.3.1
djangorestframework==3.14.0
drf-yasg==1.21.7
inflection==0.5.1
packaging==23.2
pillow==10.2.0
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
